Add a proper test for file upload and download
